Add slider function that change the grid size
i add a slider ui to the dom and add a function that change the grid
size according to its value, and i also edit the function into one big
function so event listener will be created every time the grid resized.
the reason i put all the function into one is because it is easier to
access the variables when its on the same scope.