Edit changeColor function to return real target
I edit the change color function to return the real target by using
elementFromPoint using clientX and clientY coordinate to get the current
html target. This is used due to touchmove event always returning
the same event target which is the first target that initialize the
touchmove. The downside of using elementFromPoint is reduced performance
due to The Function of DOM elementFromPoint need to loop around multiple
point of a the same element inside of all DOM tags and finally find the
elements around the points x,y.