Fix reset function and add full opacity function
i fix the reset function to create a new grid on reset instead only
changing partial style property. the create new grid is a reliable way
of making sure the whole grid style is the same as before. i also make
the reset function reset remove any added button class. the second
function i add is full opacity that will reset the opacity to null if
the opacity button is not active.