Improve pick color function and edit grid output
i improve the pick color function to toggle between opacity on or off
depending on the opacity button class, and i also separated the function
by creating 4 smaller function that will activate or inactivate the
opacity to either full or transparent. the reason i created new function
and change the pick color function is due to the function changing
opacity without the user clicking the screen if the grid is already
drawn on. We fix this by adding click event listener that will call the
start function to make sure the opacity function is only activate on
click event. i also edit the output to show the current grid size and
create the same gridSize on refresh.